Teach library to resolve string resources.
Adds a LocalizableString interface and an implementation for
String resource IDs that binds that resource ID into an oject that
later can be resolved to a localized string.

This makes it easier to pass string resources around in a type safe way,
and will later allow for binding of varargs parameters at the call site, but
string localization at the site where you have a context available, aiding
separation of concerns and lowering dependency on Android framework
classes.