Work around coverage reported as zero percent.
Seemingly the jacoco plugin does not play nice with the android gradle
library plugin. This adds report coverage for all submodules and at the
same time manages to work around the issue.

My guess is that something is funky with applying the jacoco plugin
after the library plugin, but I don't really know.