Deals with LocalizableStrings passed as varargs.
When passing varargs to the ParametrizedStringResource constructor, we
have no idea what class those objects have, but we still would like to
call LocalizableString#getString(Context context) for those objects that
implement the interface.

Using retrolambdas and functional java, it's fairly easy to map the
existing array of inparameters to a new one, and in the case of
LocalizableStrings, just call getString.

Fixes #7.