Adds StringResolver interface plus default impl.
In order to not actually leak Contexts, it is usually prudent to pass an
instance of StringResolver to presenters and models and similar,
allowing us to delegate any resolution back to e.g. the activity or
fragment, while still retaining the flexibility that passing around
LocalizableString:s gives us.

This commit also adds a default implementation that retains a context
through a weak reference, thus lowering the chance of us leaking a
context by being sloppy.